Update patch set 1
Patch Set 1: Code-Review-2

(4 comments)

Hi Sergei,

Thank you for the fix. I have +1'ed the parts I can take. But it looks like you don't have API tooling setup as the other changes breaks the API contract of the class.

That said, I am not attached to this API - but we need to change it with consultation of the community and with proper handling of the API contracts we have with our consumers.

Jonah

Patch-set: 1
Label: Code-Review=-2