Update patch set 2
Patch Set 2: Code-Review+2

(2 comments)

LGTM - the wizard just works now so I am fine with it. Slight complication is to make sure that if someone in the future creates a new class they don't create new Messages file. We'll just have to keep an eye on that.

I'm waiting for the build to confirm then I will submit.

Patch-set: 2
Label: Code-Review=+2