Update patch set 6
Patch Set 6:

(3 comments)

Hi Sergei,

We are almost there - I hope that this intensive review process on this early gerrit hasn't been offputting. But I am hoping that as we come to an agreement on how to do this subprogressmonitor -> submonitor conversion then all the future gerrits will be quick and easy to review and get submitted.

One thing that this PS introduced is a bunch of unrelated changes to comment formatting and braces. I am 100% for these changes, but please make them a new commit. Note that such commits that are purely cosmetic can (and should) be just titled Cosmetics and no bug report needed.

Thanks,
Jonah

Patch-set: 6