Update patch set 6
Patch Set 6:

> Thanks a lot, please point me what should be in a commit message
 > for each patch. Each time I little bit disappointed. Thanks.

Have a look at this: https://chris.beams.io/posts/git-commit/ It is the best write up on git commit messages - I strive for it and I am happy when I achieve it myself, all too rarely though.

HTH - your commit messages are generally fine and the info you are communicating is valuable, just a couple of times the information is ending up in the permanent record (git commit) instead of just part of the conversation of the review.

Patch-set: 6