Update patch set 3
Patch Set 3:

(7 comments)

> @Jonah I need your feedback here. I changed a bit the public API.
 > We have binary backward compatibility, but it's possible that some
 > plugin needs to change the code since I introduced a new parameter
 > for the method customizeAfterNewLineForDeclaration(). Let me know
 > what you think about it. Thanks.

That is a great question - I had a cursory look at it but I am away from my computer until Monday to look properly. So if my comment does not make sense I can do a better review later. 

This will definitely break *if* someone has their own implementation of DefaultMultilineCommentAutoEditStrategy - I have no idea if anyone actually does. The "correct" way to solve this would be a new optional interface and have the code in customizeDocumentAfterNewLine do an instanceof to know which version to call. I tried to tag lines in the review where I think this should be handled.

I thought https://wiki.eclipse.org/Evolving_Java-based_APIs had an example of this, but I can't find it - there are certainly lots of examples in the code base, in Platform if not CDT.

Patch-set: 3
Label: Code-Review=0