Update patch set 8
Patch Set 8:

> We could adjust the style based on comment prefix used by the user,
 > but it wouldn't be enough in some cases a user option is needed. I
 > could add formatter options under "comment" section but I can't see
 > a strict connection between formatter and doxygen. Is there an
 > example/internal framework to manage options from workspace or
 > project?

We should continue this discussion on the bug if possible - but short answer is I simply don't know this code at all. But as the new CDT project lead I have some obligation to learn/know areas that are not familiar to other committers. I think you are the defacto owner of this code area by being the most familiar with it and the formatter.

Patch-set: 8
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0