Bug 558827: Add build id timestamp to features properly
This was lost when the build system was moved to Maven/Tycho for
CDT 8 and restores the information.

Note that with the p2 baseline compare and replace this will show
the build id of when the branding plugin (the one with the about.mappings)
was built, which compared to the feature version may be older.

The about.mappings, as it is generated on each build, is ignored from
p2 compare and replace.

Change-Id: I38d1fcc5f618d4037baa2c8b908952421c712fc6