Update patch set 7
Patch Set 7:

The automation with script is very cool thing, but the size of PS is quite huge. I would prefer to resolve it on per-feature basis as I did for Arduino. A personal advantage for me from the "hand-made" work will be to learn code base better and to fix other branding inconsistencies. Still plan to go through our contributions after the merge of this one.

Patch-set: 7