Update patch set 7
Patch Set 7:

> Copyrights owner and year does not look consistent with the
 > enclosed content. I'm afraid it will be hard to put it in order
 > later.

Example? I think some files need to have the end date changed to 2020, but that applies to much of the code base can be fixed with the copyright tool. Note that copied files have maintained the copyright header of where they were copied from. There are a lot of files that have bad copyright headers as when they were first copied people seem to have changed the copyright headers. Most of these files are really (c) IBM 2002.

 > Also notice.html looks excessive - it just duplicates the
 > about.html Normally NOTICE.md should be created once for a
 > repository https://www.eclipse.org/projects/handbook/#legaldoc-notice

Yes, you can fix that if you want - but that is an orthogonal problem. There are lots of notice.htmls in CDT, generally one per "entry level" SDK, rather than in the root of the project. This is not a justification for keeping it that way, just I am trying to fix the branding of the features only in this PS.

 > 
 > The constructive suggestion is:
 > 1) create the PS with feature check employed (it should be small
 > one), do not merge it as it will fail.
 
I agree - the script should be its own gerrit.

 > 2) create a number of relatively small feature-focused patch sets
 > that fix the branding issues.
 > 
 > 3) merge all the PS from 2) after individual review.
 > 
 > 4) rebase PS from 1) and fix the rest of things if any: I would
 > extract "copy" functionality from there and rather focus it on real
 > checks - like existence of about.html in build.properties and so
 > on.

The copy functionality is part of the cleanliness check. If I remove it, then the cleanliness check won't report on all the errors.

 > 
 > 5) merge improved branding check.


I can split them up if that helps you review them, but I am not going to separately fix other issues at the same time - first iteration, i.e. this PS, continues the quality of the code base as it was before except it adds or fixes branding.

Patch-set: 7
Label: Code-Review=0