Update patch set 1
Patch Set 1:

> I suggest to use special releng subdir to store all the template resources.


I agree - when I originally normalized all these files I used main cdt project as the "source" for all these sorts of files so that it would be easier to edit them. But as we have the code cleanliness checks as part of the build they "can't" drift anymore.

Patch-set: 1
Label: Verified=0