Update patch set 8
Patch Set 8:

Yes, you are right, the alternative header says "Contributors to the Eclipse Foundation" but in this case you need to mention a "notice" with all the info about contributors. So, the header needs to be different in this case.

Asked Wayne about this https://bugs.eclipse.org/bugs/show_bug.cgi?id=558883

Patch-set: 8
Label: Verified=0