Update patch set 1
Patch Set 1: Code-Review-1

This looks really good - I think it is great to have.

I wasn't sure if you were looking for a review yet - but I have one bit of feedback.

The new options in the UI are appearing regardless of which documentation tool is selected. By default CDT has None and Doxygen. But as that is an extension point third-parties may have their own tools they contribute.

It may be that no one actually has one, in that case perhaps we can consider updating the GUI and deprecating the extension point with the view to removing it in CDT 10?

Perhaps a little research can find out. Maybe in the 10+ years since this feature was added everyone has settled on Doxygen?

Patch-set: 1
Label: Code-Review=-1