Update patch set 1
Patch Set 1:

A review is welcome! Actually I had the same doubt. What I could do is to enable/disable the options based on the value selected. For my knowledge the only extension about doxygen is Eclox plugin but it provides a different behavior, it doesn't extend that extension point. In my opinion the idea of this kind of extension point is not bad, but if we want to go for deprecation we can do it.

Patch-set: 1