Update patch set 1
Patch Set 1: Code-Review-1

(9 comments)

I have a couple of comments for implementation. It looks consistent with the rest of CDT source base but there are areas to improve.

As for extension point - generally I do not support its removal, but may be some polishing that we can discuss in the dedicated ticket.

Patch-set: 1
Label: Code-Review=-1