Update patch set 3
Patch Set 3:

(1 comment)

Sorry, maybe I used the wrong words. I meant that if you want a refactor or other "pieces" to clean the framework then another bug with another commit must be performed. We can't "overload" this commit just to refactor all the options management in CDT :) However the confusion for a newbie like me for CDT world is quite high about preferences because it's not really clear how to use them. For example it's not clear to me how we can use a listener if preferences has two "flavors", instance and project. Is there an automatic way for a listener to listen to everything? What happens if we were listening for instance preferences but then user enable project specific ones? In addition a listener requires a deep knowledge of what is created and what is destroyed and where, to avoid memory leak. We could use a weak reference to keep track of listeners but the WeakReference object in a set will be there anyway since a set has a strong reference to its objects. We could use a WeakHashMap or a WeakHashSet but they are not iterable and they don't really fit for a listeners container.

Patch-set: 3
Label: Verified=0