Update patch set 3
Patch Set 3:

> (9 comments)
 > 
 > I have a couple of comments for implementation. It looks consistent
 > with the rest of CDT source base but there are areas to improve.
 > 
 > As for extension point - generally I do not support its removal,
 > but may be some polishing that we can discuss in the dedicated
 > ticket.

As a general point - there is way too much public API in CDT to reasonably maintain. Just because something was once possible does not mean we should be beholden to it until the end of time. Evolving the API is as important as evolving the feature set.

Patch-set: 3
Label: Verified=0