Update patch set 1
Patch Set 1:

Its okay with me - but can you put this in org.osgi.service.prefs? It is not very CDT specific and looks genreally useful. With you being a platform committer that may be a better solution.

Note that CDT has in the past and can again track platform I-builds, so there is no delay there.

If you want to prototype / prove the implementation in an internal class here first and then move it to platform in the future that is ok with me.

Patch-set: 1
Label: Code-Review=0