Update patch set 2
Patch Set 2:

> It seems a good work. As general comment I'd like to have something
 > to install preferences listeners, maybe using a weakhashmap/weakhashset,
 > to avoid leaks, or something similar and maybe having something
 > enough smart to switch between project properties and workspace
 > properties according to what the user select, but it's another
 > story :) I think we could merge.

Thanks, Marco! This is an open question for me regarding listeners. Now we have a component that means we have activate/deactivate lifecycle hooks. So we can implement cache inside the component implementation and generally should not cache it on client side. Regarding UI that may want to listen for the storage - let's try to accumulate more implementation cases, maybe we should employ databinding first and see what will be convinient to use.

Patch-set: 2