Update patch set 3
Patch Set 3:

> Note it is not real API as class is noextend and the change is
 > private to protected not public. But feel free to provide an update
 > if you like before 9.11.0, I'm sure Torbjörn would be fine with
 > that?

Sorry for the delayed response.
Yes, I'm fine with Optional<ITool>.
I personally think it's cleaner with Optional<> in the API where the function may not return a valid object. The only reason that I did not go for the Optional<> return type was to have the smallest possible changeset while still giving sub-classes a way to fetch the ITool.

Patch-set: 3