Update patch set 1
Patch Set 1:

> Sorry for the quick merge of https://git.eclipse.org/r/#/c/156835/.
 > Thank you Alexander for taking your time to improve the API by
 > using Optional<> to indicate that it's possible that no valid ITool
 > instance can be returned.
 > May I ask why you opted for langageTool(String) instead of
 > something like getLanguageTool(String)?

I'm trying to follow OOP style for the new API if possible. The get/set approach degrades object to a data structure. We should not rape the object and extract its state with getter, but rather politely ask it with the dedicated query. You can read more here if you are interested
https://www.yegor256.com/2014/09/16/getters-and-setters-are-evil.html

Patch-set: 1