Bug 475625 - Allow user to have empty variable names
When the user selects the second wizard page was forced to use
a no empty variable name, however it's possible the parameter won't
be used and the user wants to have a matching signature. Removed
the check. The parameter name is still automatically set if the user
click directly on "Finish" button.

Change-Id: I665220b72d7b04b4bc89ffd5d764771cdfa8b243
Signed-off-by: Marco Stornelli <marco.stornelli@gmail.com>