Update patch set 1
Patch Set 1:

Thanks for the example, I have no idea about a new method to have something different with the exception of adding a new attribute as you did but I don't like it, it adds "noise" in this class and we can definitely define it as an "hack". Tomorrow I will try to think something different, but seriously, if we need something like that I prefer to revert this change waiting for 2020-06.

Patch-set: 1