Update patch set 1
Patch Set 1:

> I prefer to revert this change waiting for 2020-06.

May be not so serious :)
But I don't know if we really need this wizard to be an API. Do we know CDT extenders that needs to call the wizard programmatically or override it? I would move it to internal package at all as a part of 2020-06.

Patch-set: 1
Reviewer: Gerrit User 23745 <23745@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=0