Update patch set 1
Patch Set 1:

Yes, this should not be API going forward, at least not as it is written today. 

Marco, feel free to revert - we are hoping to solve these API issues for 2020-06 timeframe. I will be posting an initial policy document (wiki page) in the coming weeks so that we can get this resolved and announced before 2020-03 release.

Patch-set: 1