Update patch set 1
Patch Set 1:

> @Martin do you have a CQ for this contribution?

The genie didn't offer to create a CQ automatically, not sure why not in this case. Martin can't create a CQ, I can do it, but it is worth an initial review first. I will do a review in the coming days, but anyone else that can have a look would be great.

Patch-set: 1