Update patch set 5
Patch Set 5:

> Patch Set 5:
> useProjectSettings
> https://www.eclipse.org/tycho/sitedocs/tycho-compiler-plugin/compile-mojo.html#useProjectSettings

That is what we are doing. But it picks up the compiler setting from the project, ie it uses ${project.basedir}/.settings/org.eclipse.jdt.core.prefs but if there is none that is not an error. That does not pick up prefs from any parent project AFAIU

Patch-set: 5