Update patch set 5
Patch Set 5:

> > Build Failed
 > >
 > > https://ci.eclipse.org/cdt/job/cdt-verify-code-cleanliness-pipeline/843/
 > > : FAILURE
 > 
 > It says:
 > Tree is dirty - something needs to be cleaned up in your commit
 > (see above for git status/diff)
 > 
 > What does that mean? Shouldn't Jenkins do a clean checkout first?

It means that your code is not formatted to CDT standards and you are missing prefs files. The script cleans up the code base (it formats all the CDT code), and then reports an error if that actually changed anything. If you run the script locally it will leave a diff in your working copy you can amend to this commit.

Patch-set: 5
Label: Verified=0