Update patch set 8
Patch Set 8:

> About
 > the parsing strategy: actually I didn't find any test which hit the
 > added line "throw new BacktrackException();" in the method
 > simpleDeclaration, the backtrace exception is thrown before

It looks like the backtrace exception is throw from verifyDtor()?

The idea would be:

 * arrange for verifyDtor() to accept the declarator in contexts where a deduction guide could be valid
 * then, hopefully we should reach the place in simpleDeclaration() where Felix added the backtrack
 * instead of backtracking there, check for the requirements of a deduction guide, and if they are met, piece together the deduction guide from the already-parsed nodes (much liek you showed)
 * importantly, remove the "catch (BacktrackException)" clause added in declaration() which calls deductionGuide()

The idea is, right now we parse the parameter list and the return type, then throw it all away when backtracking, and parse them again in deductionGuide(). Instead, we could use the parameter list and return type we parsed the first time to create the deduction guide.

Note, this is a performance optimization. If it turns out to be tricky to implement, it's fine to land this as is and consider the optimization as a follow-up.

Patch-set: 8
Label: Verified=0