Update patch set 4
Patch Set 4:

Ok, but Jonah said it's ok and it's better to avoid any new constructor, you are saying exactly the opposite. I'm not against a rework of LookupData class but we need to do a summary :) When a class have several parameters that need to be initialized in constructor, it's better to have a builder pattern with constructor private and fields final where applicable. I could try to rework this class in this way but we need to be agree :)

Patch-set: 4
Label: Verified=0