Update patch set 4
Patch Set 4:

> Ok, but Jonah said it's ok and it's better to avoid any new
 > constructor, you are saying exactly the opposite. 

Not exactly :) I'm also not supporting the big amount of argument, and I specified there the constructor with just 2 arguments to address Jonah's concern. 

If we already have a use case to create LookupData from another LookupData - why not to just pass it to the constructor instead of unpacking its fields?

Patch-set: 4