Update patch set 7
Patch Set 7: Code-Review-2

Please do not combine refactoring into this fix. PS4 was a small/readable bit of code that was quite reviewable. There are a number of significant things that need review in the refactor. (BTW PS4 was a +6/-4 patch, PS11 is +516/-176 patch) 

I don't have an objection to making a builder class - nor to making things final - both are good improvements. However I would argue that in this case there is little value for all the extra code. LookupData is already effectively the builder class for doing a lookup and the current proposed patch simply changes a bunch of public fields (in non-API code) to getter/setter pairs. On my count only 3 of the 15 public fields became final and that took nearly 400 lines of new code. (Maybe Java will get properties one day - but this is one of the reasons that Kotlin is popular.)

Patch-set: 7
Label: Code-Review=-2