Update patch set 1
Patch Set 1:

Hi Jonah, I figured out how to make the Qt5 headers show up.  Headers such as QWidget or QApplication simply include other .h files.  The way we resolve will not figure out the content type id and so we don't create an External Translation Unit that can be used to select the C/C++ Editor for the file when we open via the Outline View.  I added some code to check for such a scenario (Include element that exists and has a WorkingCopy parent we can query for the content type id).

Patch-set: 1