Update patch set 1
Patch Set 1:

> (2 comments)
 > 
 > > Jonah, I have kept the new default device ID private for now,
 > > pending further investigation of the API tooling error. I wanted
 > to
 > > avoid the need for further API disruption until we fully
 > understand
 > > the issue.
 > 
 > > Jonah, I have kept the new default device ID private for now,
 > > pending further investigation of the API tooling error. I wanted
 > to
 > > avoid the need for further API disruption until we fully
 > understand
 > > the issue.
 > 
 > Looks good to me. I am not sure if there is any advantage to having
 > this be API anyway.


Yes, constant is not good to be an API in OOP, we may introduce a method like `Jtag#identifier` if we really need it. In general, reusable "device metadata" schema should be another outcome from the modeling effort we started for managed build. In that case the value of "identifier" will be the same for all the scenarios from project creation/import to debug configuration.

Patch-set: 1
Reviewer: Gerrit User 23745 <23745@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=0