Update patch set 8
Patch Set 8:

Yes, there's a reason actually. As first try I added Unknown with 0x80. However the return value is used as flags from PDOMName, and it there was a clash between this new flag and another flag defined in PDOMName. I mean the flag in this class are coupled with PDOMName flags, they aren't just generic flags. Adding a new flag to this class requires special attention to avoid to define something already in PDOMName. At the end I thought to use Optional to avoid this kind of problem avoiding to add new flag.

Patch-set: 8
Label: Verified=0