Update patch set 3
Patch Set 3:

> Ok, in this case we need to refactor this method, extract an helper
 > class and use the new method in cmake code and in CBuildConfiguration

I doubt that refactoring will have benefits. 
Method org.eclipse.cdt.cmake.is.core.internal.ParserDetection.expandShortFileName(String) is called in a heuristic to get a file name of an executable if none of the known names matched the leading argument in a *command-line-string*. This is a very limited use case. (I have to admit that the method name is mis-leading, but well, it's private). 

A general helper method `expandShortFileName` should have file name string as its argument, not a command-line string. Plus it should have the workaround-for-bug-in-cmake-xxx reference I added to the code  comment to give future CDT committers a chance to throw out code that is no longer needed.

Patch-set: 3