Update patch set 3
Patch Set 3:

> I doubt that refactoring will have benefits.
 > 
 > A general helper method `expandShortFileName` should have file name
 > string as its argument, not a command-line string. Plus it should
 > have the workaround-for-bug-in-cmake-xxx reference I added to the
 > code  comment to give future CDT committers a chance to throw out
 > code that is no longer needed.

I'm not suggesting to promote this method to an API. We just need to extract it, may be to package-local 
And it hsould have short unit test to demonstrate how it helps to workaround the problem, in the form of "input -> expected output". 

Currently it is not so easy to figure out why we need it here and what we are trying to workaround.

Patch-set: 3