Update patch set 1
Patch Set 1: Code-Review+1

The only minor issue I have is that this hides the AutotoolsNewProjectNature which is used by code to identify an Autotools project (e.g. CProjectBuildHelpers in Linux Tools profiling ui plug-in).  That said, the same class also uses internal classes in org.eclipse.cdt.autotools.core so it will have to use annotations to allow access so I guess it is ok.

Patch-set: 1