Update patch set 2
Patch Set 2:

Ok, thanks for the explanation, templates are hard to manage :) But at this point I'm wondering if the "final" code works since it has the same implementation. I didn't find any restriction about the use of final only on primary template, so I guess we should modify even that check in a follow-up.

Patch-set: 2
Label: Verified=0