Update patch set 3
Patch Set 3:

> @Alexander Fedorov, I would like some help writing a unit test for
 > this patch, I have effectively tested it locally, but I would not
 > want to put an 8 MB executable in the git.
 > TL;DR I would appreciate a pointer on how to unit test this
 > efficiently.

@Matthew your facilities for performance testing are great! Please submit another bugzilla ticket if you want to contribute it. I'm not sure that we need to run them for each Gerrit, but will be very nice to have it for nightly builds.

Unit testing has a bit different focus, we do not need to have huge file for it. The file can be relatively small but should contain enough data to hit all the logic of Elf class. We need it to ensure that performance improvement doesn't break the functionality. The assumption is: "slow and right code is better than fast and wrong".
Ideally, adding unit test should not require additional infra setup like "gcc" in path.

Patch-set: 3
Label: Verified=0