Update patch set 2
Patch Set 2:

Thank you Matthew for tackling this code. The Elf has evolved interestingly over many years and there are certainly many areas that you are finding for improvements.

For this bug and the rest of them in this chain - CDT's policy is:

"Non-committers must have a bug associated with every non-cosmetic commit, with bug number at the beginning of commit comment"

https://wiki.eclipse.org/CDT/policy#Committing_Code

Matthew, as a long term committer of another Eclipse project I would appreciate your input as to whether this policy should remain in place going forward. If you or Alexander want to continue the discussion as to whether bug # is needed we should continue that conversation on cdt-dev. Note - there is no requirement to have a separate bug for each gerrit. It is perfectly ok to have a single bug for cleaning up the ELF file with numerous gerrits and it makes it much easier to review how you have submitted them.

Generally for anything changing API we need a bugzilla, whether a committer or not.

Thanks again Matthew!

Patch-set: 2
Label: Code-Review=0