Update patch set 4
Patch Set 4:

> > I will abandon this gerrit, please do not post commants any
 > longer
 > > here.
 > > I just need to keep this for a while to have number of changed
 > > lines.
 > 
 > :-( sorry that doesn't work. Each of the split pieces will then
 > need a CQ. You are on the committer track - once that happens you
 > will be LOC free.
 > 
 > There has just been a CQ change that I didn't take advantage of on
 > your first contribution that means we can megre before legal review
 > is complete.

I should clarify, you are free to separate the Gerrits, but it is not the strict LOC that matters for CQ processing, but the amount of new IP. E.g. If you are refactoring something that means there is lots of lines of code changed, but not much new IP then a CQ is not needed, no matter how many lines of code. Similarly, a very small change may need a CQ if it looks like a substantial value of IP. 

HTH - and finally that is all off topic if your motivation is not CQ related.

Patch-set: 4