Update patch set 4
Patch Set 4:

It is OK to split change for code reviewing reasons - it cannot be used as a way to skirt around the 1000 LOC IP review.

The 1000 LOC is a completely arbitrary line in the sand as to when code needs to go through additional checks. It is a proxy for when a significantly large IP contribution is being made. So having 1 commit or 6 commits does not change that. Indeed if a committer is concerned that a smaller LOC contribution may be significantly large in terms of IP there needs to be a CQ too.


-----

So for the way forward here: if this change is not new IP is to just submit the changes in whatever way makes sense from a code review standpoint. If a contribution is of 1000 LOC and you are sure it is OK, make a comment in the gerrit saying so before submitting so the decision is recorded. For example, just renaming classes and many refactoring of some code can generate large diff counts.

However if this change is new IP, then it just needs to be submitted as a CQ. That should not slow things down too much as we are eligible for "parallel IP" where code can be checked in very early in the review process.


-----

Finally, Martin as you familiarise yourself with Eclipse processes, procedures and continue to be involved in the project you will soon be a committer if you want. At that point none of your code needs to go through the extra checks as the committer agreement covers IP issues.


I hope that helps.

Patch-set: 4