Update patch set 4
Patch Set 4:

> >
 > > It was Alexander who suggested to split this and to avoid another
 > > CQ [1] [2].
 > >
 > 
 > Yes, it was me. Because of two reasons.
 > 1) I would like to see shorter and more focused change sets. Thanks
 > you for doing extra work to split your initial commit, Martin.

+1

 > 2) I have an experience of using ">1000 LOC" rule in the platform
 > that can be hardly explained by IP behind it. Please have a look at
 > discussion here: https://git.eclipse.org/r/#/c/149953/7 . The
 > change was to add `|| resolvedURL.startsWith("https://")` and to
 > provide a short unit test for it.
 > 
 > PS I'm not a lawyer and hopefully will never be :)

+1 - that is why it needed a CQ. 

I have had explicit confirmation on the past that the LOCs that are things like prefs files do not need CQ. 

I'll submit this change as CQ and hopefully we can just get it approved and have that approval applied to modifications if the code so we don't have to submit this Gerrit. 

---

Martin, thanks for your patience as we figure out what to do here.

Patch-set: 4