Bug 560554. Launch target not populated into defaultdescriptors.
Make explicit the decisions to push down to providers to
populate configuration by checking if descriptor supports target. Make
explicit the delete/reset logic by a new method on the interface, as
opposed to checking if the descriptor is an instanceof
DefaultDescriptor. This enables descriptors which support targets, and
can be deleted.

Change-Id: Id4e3126a70d8cb091d4520acf5d21c5205e0c745
Signed-off-by: Tad Adams <tadams@blackberry.com>