Update patch set 3
Patch Set 3:

> (4 comments)
 > 
 > Not sure why code cleaness step is failing now, just has this in
 > the log:
 > Tree is dirty - something needs to be cleaned up in your commit
 > (see above for git status/diff)
 > 
 > Any ideas?

You have fixed it in PS3 I think, the full output from the log is:

09:55:50  diff --git a/launchbar/org.eclipse.launchbar.core/src/org/eclipse/launchbar/core/ILaunchDescriptorType.java b/launchbar/org.eclipse.launchbar.core/src/org/eclipse/launchbar/core/ILaunchDescriptorType.java
09:55:50  index 1b920d6e66..a03a5ede07 100644
09:55:50  --- a/launchbar/org.eclipse.launchbar.core/src/org/eclipse/launchbar/core/ILaunchDescriptorType.java
09:55:50  +++ b/launchbar/org.eclipse.launchbar.core/src/org/eclipse/launchbar/core/ILaunchDescriptorType.java
09:55:50  @@ -52,5 +52,4 @@ public interface ILaunchDescriptorType {
09:55:50   		return true;
09:55:50   	}
09:55:50   
09:55:50  -
09:55:50   }
09:55:50  Tree is dirty - something needs to be cleaned up in your commit (see above for git status/diff)

which is showing that you need to remove the line with the "-" to get your file formatted properly. I am not sure how you are editing the file as the save actions in the Eclipse IDE should be applying these rules.

I will try to add some more info next to "something" for future contributors to understand better.

Patch-set: 3