Update patch set 4
Patch Set 4:

The save actions are set per project so the workspace settings is not actually used for this. It makes setup easier, but clearly there is something wrong.... 

I just tested in 2020-03 release and it removed the line with the save actions when I modified a line. If you can reproduce a problem with the save actions I would be very interested to hear.

Thanks!

Patch-set: 4