Bug 562181: Fix auto indent after constructor with access specifiers
Some heuristics in skipToStatementStart require knowledge of previous
token. It is initialized with value from a fToken, but it current value
may be invalid because looksLikeMethodDecl changes fPossition value
without adjusting a fToken. Using invalid value may lead to false
triggering of heuristics. Restore fToken value when change fPossition in
looksLikeMethodDecl to prevent this.

Change-Id: If0c2c0577c89a983e1479587409f293c3d0db7be
Signed-off-by: Andrey Mozzhuhin <amozzhuhin@yandex.ru>