Update patch set 11
Patch Set 11:

> >
 > > (Off topic...)
 > >
 > > Where do you expect this error to come out? Isn't it the same as
 > > all other Maven projects, or does CDT have hidden somewhere
 > within
 > > its massive poms something that is making us act in a
 > non-standard
 > > way? If the latter, we should fix that so people aren't confused
 > in
 > > the future.
 > 
 > Ideally like the following:
 > https://ci.eclipse.org/cdt/job/cdt-verify-code-cleanliness-pipeline/1135/
 > : FAILURE. Compilation failed.
 > 
 > and then all the rest regarding console output and documentation
 > 
 > The strategic goal is to give enough info regarding failure without
 > visiting Jenkins.

OK. I don't think I will have time to do any of that - I can't justify the time to implement that when it is all available behind a click. I encourage you to explore what functionality the gerrit plug-in has. IIRC you can send some info to a file and send that file contents to gerrit. 

(BTW we should move this OT discussion if it continues out of Matthew's gerrit ;-)

Patch-set: 11
Label: Verified=0