Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (8 comments)
> 
> @Jeff The approach used in the code should be replaced with org.eclipse.core.runtime.jobs API:
> * Job#belongsTo(Object family)
> * IJobManager#join(Object family, IProgressMonitor monitor)
> and others regarding IShedulingRule and IJobListener

Hi Alexander.  I am trying to avoid join().  There was a deadlock created by the rechecking of image targets so I created jobs to allow the init and changeEvent methods to return.  A spurious problem has been reported once in a while and appears to be impossible (list of targets has a null target in it which should never occur).  I believe this may be caused by the check jobs not being complete and other code accessing the target list.

Add to that, we are doing a point release for 2020-06 so I can't add a new API.  Using the getStatus() call seemed a way to kludge this, however, it is called elsewhere so I was avoiding using a join() in case I caused the deadlock again (much more serious issue).  Anyway, that is the explanation of why I am doing this so weirdly.

Patch-set: 1