Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

I don't like this change. The code used to fit in a single screenful, now it is separated in three separate files.

Note that I have only scored it -1, not -2. If you are seeing a value to this change for end users, CDT developers or CDT extenders, then you are free to submit it. If this code was in a part of CDT we were not actively trying to move away from, that may also have changed my score and the amount of time I spend reviewing it.

Patch-set: 2
Label: Code-Review=-1