Update patch set 2
Patch Set 2:

> (1 comment)
 > 
 > I don't like this change. The code used to fit in a single
 > screenful, now it is separated in three separate files.
 > 

Yes, the style is different from surrounding one, but for me it looks better as each of these separate files has high cohesion and easily testable. Because of that I beleive it is more maintainable than adding more protected methods to a deep hierarchy.

 > Note that I have only scored it -1, not -2. If you are seeing a
 > value to this change for end users, CDT developers or CDT
 > extenders, then you are free to submit it. If this code was in a
 > part of CDT we were not actively trying to move away from, that may
 > also have changed my score and the amount of time I spend reviewing
 > it.

Well, I do not really insist on this very change. May be it is wrong package and bundle for extracted functionality. I would say it was a test change to see if OOP style can be introduced to the exisiting source base.

Patch-set: 2
Label: Verified=0